+++ CHANGELOG.md
@@ -21,10 +21,16 @@ Adding a requirement of a major version of a dependency is breaking a contract.
 Dropping a requirement of a major version of a dependency is a new contract.
 
 ## [Unreleased]
-[Unreleased]: https://github.com/atlassian/virtual-users/compare/release-3.10.0...master
+[Unreleased]: https://github.com/atlassian/virtual-users/compare/release-3.11.0...master
+
+## [3.11.0] - 2020-03-04
+[3.11.0]: https://github.com/atlassian/virtual-users/compare/release-3.10.0...release-3.11.0
 
 ### Added
 - Generate realistic user names in `RestUserGenerator`.
+
+### Fixed
+- Harden HTML page source dump generation.
 
 ## [3.10.0] - 2019-08-02
 [3.10.0]: https://github.com/atlassian/virtual-users/compare/release-3.9.1...release-3.10.0